+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * 【core  】     改造IndexedComparator，增加InstanceComparator
 * 【extra 】     增加CglibUtil
 * 【extra 】     增加Ipv4Util（pr#161@Gitee）
+* 【extra 】     增加CalendarUtil和DateUtil增加isSameMonth方法（pr#161@Gitee）
 
 ### Bug修复#
 * 【poi   】     修复ExcelBase.isXlsx方法判断问题（issue#I1S502@Gitee）